requirements.txt
# packages in environment at RL:
#
# Name                    Version                   Build  Channel
blas                      1.0                         mkl  
brotli                    1.0.9                h2bbff1b_7  
brotli-bin                1.0.9                h2bbff1b_7  
ca-certificates           2023.12.12           haa95532_0  
cffi                      1.16.0           py39h2bbff1b_0  
colorama                  0.4.6            py39haa95532_0  
console_shortcut          0.1.1                         4  
contourpy                 1.2.0            py39h59b6b97_0  
cycler                    0.11.0             pyhd3eb1b0_0  
eigen                     3.4.0                h59b6b97_0  
ffmpeg                    4.2.2                he774522_0  
filelock                  3.13.1           py39haa95532_0  
fonttools                 4.25.0             pyhd3eb1b0_0  
freetype                  2.12.1               ha860e81_0  
fsspec                    2023.10.0        py39haa95532_0  
giflib                    5.2.1                h8cc25b3_3  
glib                      2.69.1               h5dc1a3c_2  
gmpy2                     2.1.2            py39h7f96b67_0  
gst-plugins-base          1.18.5               h9e645db_0  
gstreamer                 1.18.5               hd78058f_0  
hdf5                      1.12.1               h51c971a_3  
icc_rt                    2022.1.0             h6049295_2  
icu                       58.2                 ha925a31_3  
importlib_resources       6.1.0            py39haa95532_0  
intel-openmp              2023.1.0         h59b6b97_46320  
jinja2                    3.1.2            py39haa95532_0  
jpeg                      9e                   h2bbff1b_1  
kiwisolver                1.4.4            py39hd77b12b_0  
krb5                      1.20.1               h5b6d351_0  
lerc                      3.0                  hd77b12b_0  
libbrotlicommon           1.0.9                h2bbff1b_7  
libbrotlidec              1.0.9                h2bbff1b_7  
libbrotlienc              1.0.9                h2bbff1b_7  
libclang                  14.0.6          default_hb5a9fac_1  
libclang13                14.0.6          default_h8e68704_1  
libdeflate                1.17                 h2bbff1b_1  
libffi                    3.4.4                hd77b12b_0  
libiconv                  1.16                 h2bbff1b_2  
libogg                    1.3.5                h2bbff1b_1  
libpng                    1.6.39               h8cc25b3_0  
libpq                     12.15                h906ac69_1  
libprotobuf               3.20.3               h23ce68f_0  
libtiff                   4.5.1                hd77b12b_0  
libuv                     1.44.2               h2bbff1b_0  
libvorbis                 1.3.7                he774522_0  
libwebp                   1.3.2                hbc33d0d_0  
libwebp-base              1.3.2                h2bbff1b_0  
libxml2                   2.10.4               h0ad7f3c_1  
libxslt                   1.1.37               h2bbff1b_1  
lz4-c                     1.9.4                h2bbff1b_0  
markupsafe                2.1.1            py39h2bbff1b_0  
matplotlib                3.8.0            py39haa95532_0  
matplotlib-base           3.8.0            py39h4ed8f06_0  
mkl                       2023.1.0         h6b88ed4_46358  
mkl-service               2.4.0            py39h2bbff1b_1  
mkl_fft                   1.3.8            py39h2bbff1b_0  
mkl_random                1.2.4            py39h59b6b97_0  
mpc                       1.1.0                h7edee0f_1  
mpfr                      4.0.2                h62dcd97_1  
mpir                      3.0.0                hec2e145_1  
mpmath                    1.3.0            py39haa95532_0  
munkres                   1.1.4                      py_0  
networkx                  3.1              py39haa95532_0  
ninja                     1.10.2               haa95532_5  
ninja-base                1.10.2               h6d14046_5  
numpy                     1.26.2           py39h055cbcc_0  
numpy-base                1.26.2           py39h65a83cf_0  
opencv                    4.6.0            py39ha36de5b_5  
openjpeg                  2.4.0                h4fc8c34_0  
openssl                   3.0.12               h2bbff1b_0  
packaging                 23.1             py39haa95532_0  
pcre                      8.45                 hd77b12b_0  
pillow                    10.0.1           py39h045eedc_0  
pip                       23.3.1           py39haa95532_0  
ply                       3.11             py39haa95532_0  
pycparser                 2.21               pyhd3eb1b0_0  
pyparsing                 3.0.9            py39haa95532_0  
pyqt                      5.15.10          py39hd77b12b_0  
pyqt5-sip                 12.13.0          py39h2bbff1b_0  
python                    3.9.18               h1aa4202_0  
python-dateutil           2.8.2              pyhd3eb1b0_0  
pytorch                   2.1.0           cpu_py39hb0bdfb8_0  
qt-main                   5.15.2               h879a1e9_9  
qt-webengine              5.15.9               h5bd16bc_7  
qtwebkit                  5.212                h2bbfb41_5  
setuptools                68.0.0           py39haa95532_0  
sip                       6.7.12           py39hd77b12b_0  
six                       1.16.0             pyhd3eb1b0_1  
sqlite                    3.41.2               h2bbff1b_0  
sympy                     1.12             py39haa95532_0  
tbb                       2021.8.0             h59b6b97_0  
tk                        8.6.12               h2bbff1b_0  
tomli                     2.0.1            py39haa95532_0  
tornado                   6.3.3            py39h2bbff1b_0  
tqdm                      4.65.0           py39hd4e2768_0  
typing-extensions         4.7.1            py39haa95532_0  
typing_extensions         4.7.1            py39haa95532_0  
tzdata                    2023c                h04d1e81_0  
vc                        14.2                 h21ff451_1  
vs2015_runtime            14.27.29016          h5e58377_2  
wheel                     0.41.2           py39haa95532_0  
xz                        5.4.2                h8cc25b3_0  
zipp                      3.11.0           py39haa95532_0  
zlib                      1.2.13               h8cc25b3_0  
zstd                      1.5.5                hd43e919_0  
